+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Changelog for Cass Operator, new PRs should update the `main / unreleased` secti
 
 ## unreleased
 
-* [BUGFIX] [#930](https://github.com/k8ssandra/cass-operator/issues/930) Add the resource-hash annotation to the NodePort service so the operator reconciles updates to it; previously the NodePort service was only created and never updated, so changes to additionalServiceConfig.nodePortService were ignored.
+* [BUGFIX] [#930](https://github.com/k8ssandra/cass-operator/issues/930) Add the missing resource-hash annotation to the NodePort service so changes to it are detected and reconciled.
 
 ## v1.31.0